+++ actix-service/src/macros.rs
@@ -147,6 +147,7 @@ mod tests {
 
         forward_ready!(inner);
 
+        #[allow(clippy::unit_arg)]
         fn call(&self, req: ()) -> Self::Future {
             self.inner.call(req)
         }