+++ actix-utils/src/lib.rs
@@ -6,10 +6,5 @@
 #![doc(html_favicon_url = "https://actix.rs/favicon.ico")]
 
 pub mod counter;
-pub mod dispatcher;
-pub mod mpsc;
-mod poll_fn;
-pub mod task;
+pub mod future;
 pub mod timeout;
-
-use self::poll_fn::poll_fn;