+++ local-waker/CHANGES.md
@@ -3,5 +3,5 @@
 ## Unreleased - 2021-xx-xx
 
 
-## 0.1.0 - 2021-03-29
+## 0.1.1 - 2021-03-29
 * Move `LocalWaker` to it's own crate.